retree: rename to satellite
