+++ config.json
@@ -1423,7 +1423,7 @@
       ]
     },
     {
-      "slug": "retree",
+      "slug": "satellite",
       "uuid": "18467e72-e3b0-403c-896a-ce337f96e6f7",
       "core": false,
       "unlocked_by": "markdown",
