+++ exercises/satellite/example.py
