+++ exercises/satellite/satellite.py
